backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # Removed - not needed for asyncpg
asyncpg==0.29.0

# Cache & Queue
redis==5.0.1
celery==5.3.6
kombu==5.3.5

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
aiohttp==3.9.1
httpx==0.26.0

# Notifications
twilio==8.11.1
resend==0.8.0

# NLP
spacy==3.7.2

# Utilities
python-dotenv==1.0.0
pytz==2024.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0